# Make interface generic, Drop StaticArrays as hard dependency