--- a/Project.toml
+++ b/Project.toml
@@ -2,18 +2,15 @@ name = "Contour"
 uuid = "d38c429a-6771-53c6-b99e-75d170b6e991"
 version = "0.5.7"
 
-[deps]
-StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
-
 [compat]
-StaticArrays = "0.10,0.11,0.12, 1.0"
 julia = "0.7, 1"
 
 [extras]
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 OffsetArrays = "6fe1bfb0-de20-5000-8ca7-80f57d26f881"
+StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["LinearAlgebra", "StatsBase", "Test", "OffsetArrays"]
+test = ["LinearAlgebra", "StatsBase", "Test", "OffsetArrays", "StaticArrays"]