--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -1,9 +1,10 @@
 using Contour, Test
 
+@test length(detect_ambiguities(Base, Contour)) == 0
+
 include("verify_vertices.jl")
 include("interface.jl")
 
-# @show detect_ambiguities(Base, Contour) # should be zero but there are a few stragglers in the imports
 
 #issue 59
 @inferred collect(())