@sjkelly 2022-03-24
I am not 100% this is the API we want yet, but it is a step in the right direction for a point release.